WebExtensions: Redesign Browser Actions to have a flexible presentation
Previously the WebExtensionManager would directly poke around inserting
widgets into the HeaderBar. This falls apart with adaptive views so
switching to mobile would lose access to extensions.

This changes the design where WebExtensionsManager exposes a list of
EphyBrowserAction objects and the UI tracks this list and decides
how to show them in any mode or in multiple windows.

As well as this change the UI was changed so now all browser actions
are shown in a popover and the popover doubles as the browser actions
popup UI when clicked. This matches Firefox's UI (but is even smoother!).

The ability to show all actions as a dedicated button in the headerbar
was lost for now. I think this is ok for now but ideally we implement
a solution similar to other browsers where the user can select which
ones to be dedicated buttons and which ones to go into the dropdown.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1199>