lib/widgets/: Move ephy-file-chooser.* to lib/
The goal here is to get rid of the confusing and annoying lib/widgets/
layer. ephy-file-chooser.* can't be moved to neither embed/ nor src/
because both of them use it, but the file chooser is not more specific
to the web view (i.e. embed/) than it is to the main src/ code: both of
them use its utility function. Even if src/ does depend on embed/, it
does not make sense to put the file chooser code under embed/, so even
if not ideal, the code will live under lib/ even if it is currently
almost exclusively non-GTK stuff (apart from e.g. ephy-gui.*).

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1208>