location-controller: use correct tab view when closing overview tab
When an ephy-tab:// URI is used to activate a tab in a different window,
the code switches the tab_view variable to point to a different window's
tab view, then forgets that it did so and attempts to use it as if it
were for the current window in order to close the current tab. Fix this
by resetting the tab_view variable to point to the original window's tab
view.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1220>