window: Disable key event handling
This is a temporary kludge to dodge #1915. I don't think we actually
need this code anymore because I'm not aware of any practical cases
where Epiphany accelerators conflict with websites. The most important
motivating cases were previously Ctrl+B, which used to open the
bookmarks dialog, and Ctrl+I, which used to open an incognito window.
But these accelerators no longer exist.

However, it's true that we probably do want a way to decide whether the
key events go to the window or the web view, so Alexander has requested
that this code not be completely removed.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1227>