Replace "web application" with "web app"
menu but the help, keyboard shortcuts and about:applications still have
the previous wording. This change updates those to use the same wording
as in the page menu.

See also the initiative to replace "application" with "app":
https://gitlab.gnome.org/Teams/Design/initiatives/-/issues/123

The notes "It is not possible to create/delete Web Apps if you are using
Flatpak to run Web" are removed from the help pages as these are
possible.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1237>