Remove ephy_gui_ensure_window_group()
Coverity discovered a use-after-free in ephy_gui_ensure_window_group(),
where the newly-created GtkWindowGroup is freed immediately before it is
returned. This could only possibly make sense if GtkWindowGroup has very
weird refcounting semantics, but there is no indication of this in its
documentation. Fortunately, the buggy code is unreachable because it
only occurs if gtk_window_get_group() returns NULL, which is impossible
because the function is not nullable. It will always return the default
window group, not NULL. This means the entire condition can be removed,
so ephy_gui_ensure_window_group() can be simplified to just return
gtk_window_get_group() without changing behavior. But none of the
callers actually look at the return value. And gtk_window_get_group()
has no side-effects, so it does not "ensure" anything. Therefore all of
this code is doing nothing, and can be safely removed.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1256>