add-bookmark-popover: fix refcounting error in popover_shown()
We use bookmark later on after the ref has already been dropped, which
is no good. Found by Coverity.

This should not be a problem in practice because the
EphyBookmarksManager will keep the bookmark alive, but it's not good to
rely on this. We should have our own ref for as long as we're using the
bookmark.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1257>