context-menu-commands: improve set_image_as_background()
Let's use g_autofree and g_autoptr here. This notably avoids use of the
EphyDownload after we have dropped our ref on it, which Coverity doesn't
like. It's not a problem in practice because the EphyDownloadsManager
will still have a ref, but we should fix it anyway.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1257>