file-dialog-utils: remove ephy_file_dialog_add_shortcuts()
The underlying API was just removed in gtk!5396, so we have to stop
using it. This is fine, though, because all it does is add the downloads
folder to the file chooser sidebar, which is redundant because XDG
directories are already there by default for a decade or so now. So this
code is not useful and can disappear.

Thanks to Alexander for telling me exactly what to do here.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1261>