suggestion-model: further improve robustness of get_item()
Company pointed out that we should probably handle all possible invalid
values here instead of allowing them to throw criticals inside the call
to g_sequence_get(). An easy way to do that is to check the iterator
before using it.

This is a follow-up to !1289.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1290>