Make EphySearchProvider a subclass of EphyEmbedShell
EphySuggestionModel depends on EphyEmbedShell, which seems like a pretty
reasonable thing to do. Prior to 44 it only depended on the *existence*
of the EphyEmbedShell, which is created by the search provider. But
since 44 it additionally depends on the embed shell *running* because
the favicon database is expected to exist but it's no longer created
unless the embed shell is actually started.

One way to resolve this would be to enable favicons when the embed shell
is created rather than when it is started, but it's probably more robust
to just turn the search provider into an embed shell instance.
Hopefully.

Fixes #2006

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1303>