Delay EphyWindow destruction when discarding modified forms
GtkWindow will crash in dispose when trying to unset its transient-for
parent window if the parent window is already destroyed. That is, we
need to ensure that child windows get destroyed before their parents do.
In this case, the child AdwMessageDialog is displaying a button that
lets you destroy its parent window. Instead of doing that immediately,
chill out and schedule it to happen on the next idle main context
iteration.

Fixes #2043

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1315>