Move Sync Now button to sync dialog header bar
The Sync Now button is currently broken because the AdwComboRow consumes
the click and opens the combo selector instead, so we need to move it
somewhere else. Easiest place to put it is the header bar.

This is technically a UI freeze break, but it's consistent with the
*spirit* of UI freeze in that it's the simplest way to fix the bug.
Alice says finding a way to get the button to consume the click while
it's a child of the AdwComboRow is not planned.

Fixes #2042