permissions-manager: add asserts regarding storage access permission requests
These permissions _could_ be stored in the permissions manager, but
they're not. And it will crash if you try to do it, so let's crash a
little bit sooner with an assert to document that it's not intended to
be done.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1320>