preferences: Add new WebKit features page
Add a new preferences page which allows toggling development and
experimental WebKit features, using new API [1] which available from
2.41.4 onwards. The features are grouped by their category (using
AdwPreferencesGroup) and elements (using AdwActionRow) have both a
switch to enable/disable them, as well as a button to reset the feature
to its default setting, and a label indicating the status (maturity)
of the feature.

The page is always enabled in Tech Preview builds, and gated behind
a gsetting otherwise, so it can be enabled also for regular builds
if so desired.

[1] https://bugs.webkit.org/show_bug.cgi?id=255518

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1323>