permissions-manager: clarify which permissions are stored
Most permissions get remembered in the permissions store, but the
storage access ("cookies") permission gets requested each time, and I'm
about to add a second permission type (clipboard access) that behaves
the same, and webcam/microphone is also a bit simple.

Note this incidentally fixes some incorrect handling of the cookies
permission in ephy-window.c, where when we're dealing with a cookies
permission we execute code that assumes webcam/microphone permission.