Fall back to favicon when failing to set web app icon
On the web app creation dialog, if we fail to create a texture for the
web app icon, let's fall back to using a favicon instead rather than
unnecessarily give up entirely.

Also, improve the error reporting for good measure.

Fixes #1989 sort of. This will ensure we don't totally fail to create
the web app when the image format is unrecognized. That will be fixed
next.