Fix the Mute Tab context menu item
Currently the state we display does not match the actual state of the
tab. We need to update the state when opening the context menu based on
the tab that the user has clicked on. Since we currently fail to do
that, we are instead showing the state of whichever tab was muted or
unmuted most recently rather than the state of the tab being clicked on.

This code is a little hacky as we are assigning the state of a single
tab into global state, but this is not really any worse than setting
global actions to be enabled/disabled based on the state of a tab, so
it's no worse than the existing surrounding code.

https://bugs.webkit.org/show_bug.cgi?id=253375

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1354>