gsb-storage: unset self->is_operable more carefully
self->is_operable is supposed to indicate whether the database is in a
usable state. ephy_gsb_storage_clear_db() is forgetting to do so. It's
followed by a call to ephy_gsb_storage_init_db(), but this, too, can
forget to do so. The rest of the code assumes the invariant self->db ==
self->is_operable and will crash trying to dereference self->db if
self->is_operable is incorrect.

Fixes #2132

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1366>