tab-view: fix self-reference preventing tab view destruction
The EphyTabView contains an AdwTabView. EphySession needs to be able
to get the EphyTabView corresponding to an AdwTabView, so EphyTabView
stores itself into the AdwTabView using g_object_set_data_full().
Problem is it takes an unnecessary reference to itself, therefore
guaranteeing it can never die. Break this.

Most notably, this was causing all remaining EphyWebViews to be leaked
when the window is closed. Users would not notice when closing the final
window, but when closing any other window it was possible to notice
audio still playing forever during media playback, for example.

This is actually a case where inheritance would work better than
composition. Using g_object_set_data() to "cast" between the parent and
child is awful. But I'm not sure how to improve this without making
AdwTabView derivable.

Fixes #2130

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1367>