window: don't dismiss password popover when loading page
Since !1337, when loading a new page in the web view, all permission
request popovers are dismissed. This is correct for standard permission
request popovers that are backed by WebKit permission requests, but it
doesn't make sense for password popovers. It results in us dismissing
the Save Password? request almost immediately after it is presented to
the user. Instead, keep it around indefinitely.

Fixes #2113

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1368>