Be smarter about reading form action
In JavaScript, the action property of a HTMLFormElement normally refers
to the action attribute of the element. However, if the form contains
any elements with name "action" then it refers to that element instead.
Oops. That's not what we want the password manager to see.

Fixes #2114

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1370>