bookmark-row: implement ListBoxRow.activate (enter)
implement the activate method, which gets called when enter is pressed. the click gesture handler is still there to handle middle click and modifiers.
shortcuts such as ctrl+enter to open in a new tab are not implemented in this commit
also the function to open the bookmark of the row is moved from BookmarksPopover to BookmarkRow, because it's needed there

fixes a part of #2145

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1374>