ephy-window: Ctrl-Shift-o now toggles overview
Pressing Ctrl-Shift-o opens the tab overview.
However, pressing it again while inside the overview does not close it.
This is inconsistent and feels unnatural.

This fix changes the ephy_window_open_tab_overview() function to
ephy_window_toggle_tab_overview(). It now acts as a toggle, so it can
both close and open the overview.

This does cause a problem when holding down Ctrl-Shift-o (the screen
flashes between overview open and closed), but it could
be easily resolved by mapping toggling the overview to when the
keyboard shortcut is released, not right when it is pressed.

Fixes: #2147
Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1376>