flatpak-utils: don't force OpenURI portal to display its UI
By default, the OpenURI portal will ask the user to select an
application to open the link the first few times, but will eventually
stop doing so and just open the link automatically. This is what users
what, but we're passing the XDP_OPEN_URI_FLAG_ASK flag to suppress this
behavior and always show the UI. Since this is annoying users and
doesn't have any particular benefit, let's remove it.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1387>