Don't rely on indexing when deleting bookmark tags
As a performance optimization, EphyBookmarksManager currently provides
the index of the tag in its tags list to help EphyBookmarksPopover
decide which GtkListBoxRow to remove. But this is wrong because the
order of rows may not match the EphyBookmarksManager's order of tags.
For example, when removing the tag from its last bookmark without
deleting the tag, we remove the tag row from the list box, but the tag
still exists. The Favorites tag also always exists even if it's not
displayed due to lack of associated bookmarks. (I was also briefly
concerned that the list box contains both bookmarks and tags, not just
tags alone, but this is fine because the bookmarks always sort after the
tags.) So we cannot establish any correspondence between the position of
the tags in the EphyBookmarksManager vs. the position of the tags in the
list box.

Fix this by just searching for the right tag. It's not like deleting a
bookmark tag needs to be optimized for speed. This happens quite rarely
and will be good enough unless the user has millions of bookmarks.

This fixes a critical when deleting bookmark tags.

Fixes #2180

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1388>


(cherry picked from commit a747215ca08d1d1a93af634b159738246701fb7f)