Remove Safe Browsing support
This is unfortunate, but this feature has been disabled everywhere
except Epiphany Tech Preview due to changes in Google's terms of
service. We have to keep Google API keys secret now, and it's just not
possible unless build secrets are allowed. We can do that in Tech
Preview, but Linux distros cannot do this. Keeping the feature around
only for ourselves and not for normal users is just not very useful.

Closes #1393
Closes #1334
Closes #1699
Closes #1721

Related: #682