web-view: register ucm handlers in constructed
Best practice is to do the bare minimum work in new, so there is less to
do when using g_object_new() directly. We don't ever do that in
Epiphany, but it's still nice to follow best practices, and also avoids
a small amount of code duplication.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1409>