Fix EphyWindow leak when opening install web application dialog
We are leaking the XdpParent here, which refs the EphyWindow.

This sort of fixes #2214. I say "sort of" because any EphyWindow leak
anywhere will trigger this assertion, and it's not particularly likely
that this leak is the one discovered by Kdwk. But it's one such leak
that I know how to reproduce. Please continue to report issues for
additional leaks, especially if you find a way to reproduce it.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1410>