Fix the user help
The user help does not lauch successfully since switching to
GtkUriLauncher. Problem is it is going through the OpenURI portal, so
yelp gets opened on the host system instead of in the Epiphany flatpak
where the user help is installed. We need to make sure to open yelp
inside the flatpak.

This avoids gtk#6135