Fix more memory corruption caused by ephy_string_shorten()
This reminds me of !1106. This function is supposed to return a
newly-allocated string. Failure to do so results in a use after free
later on. This *might* be responsible for #2021 (probably not, but
maybe).

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1423>


(cherry picked from commit 8717047c12f5721d1da9b7042284ffdf68b461a4)