sync-service: switch to SecretRetrievable API
secret_password_search_finish() returns a list of SecretRetrievable
objects. We incorrectly assume it returns a list of SecretItems.
SecretItem seems to be the only public implementation of
SecretRetrievable, but clearly it's not the only implementation.

We get to the same end result -- a SecretValue object -- either way, but
SecretRetrievable has an async interface, so the code is a bit more
complicated now. It's surely for the better.

Fixes #1361 (probably) (hopefully :)

This backport is modified to not change translatable strings.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1462>