session: move and update comment regarding session sanity check
We had a comment explaining what is going on here, but it's where the
function is called rather than where it's implemented, which isn't a
great place for it. So I wasted my time writing a new version of the
comment. Move and update it.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1470>