session: don't save non-openable URLs
It's possible for JavaScript to open a tab with an empty URL or a data
or blob URL. It's not possible to restore such a tab from session state
because WebKitGTK reloads the entire page (unlike WebKit on macOS). So
don't save them at all.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1470>