file-helpers: allow overriding MIME type when launching file
Fixes #2301, hopefully. We are failing to launch these files when their
MIME types are detected as application/x-zerosize. That's unhelpful. We
already know what MIME type we should use, so no need to detect it from
the file contents.

This backport for the gnome-46 branch makes fewer changes to the code
than the master branch version.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1485>