notification-container: don't crash on duplicate notifications
The code that checks for duplicate notification attempts to remove the
newly-created notification from the container by mistake. It hasn't been
added to the widget hierarchy yet. Maybe the code here was trying to
remove the original notification, but it's probably best to just discard
the new one.

Fixes #2344

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1487>