Normalize placeholder URL
When the session is saved, "ephy-about:overview" URL is substituted with
"about:overview". When the session is later restored and delay loading
is used, `ephy_web_view_set_placeholder()` is called to show an empty
page until the requested page is fully loaded.

But since `ephy_web_view_set_placeholder()` does not normalize the
passed URL and "about:overview" is not listed in `do_not_show_address`,
it appears in the URL bar as is.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1492>