Fix handling of moved form elements in form-controls-associated handler
This could fix the password manager on potentially many websites. The
problem is we currently process forms only when the form is created. If
a password element is later added to the form, we miss it.

There are a bunch of code changes here, but the premise is simple:
check all the DOM elements we receive, find each one's parent form
element, and if it hasn't already been handled in the current emission
of form-controls-associated, then handle it again.

This notably allows the password manager to find the password field on
https://auth.max.com/create-account which it previously missed.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1495>