Generate secure password using libpwquality
This adds detection of "new password" elements (i.e. password elements
that are for entering a brand new password rather than an existing one).
Display an option for the user to generate a secure password. It's
intended to be stored in the password manager.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/2108>