Respect new-password autocomplete hint when prompting to save a password
When we are not autocompleting a previously-saved password, we currently
have a complex heuristic to guess which password field we should use
when proposing to save the user's password. This actually works pretty
well, but if the website is kind enough to tell us which field to use,
then we should just take that instead.

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/password#allowing_autocomplete

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1495>


(cherry picked from commit 4ac9b085330721c54d7e291f6f925466f96846f4)

Co-authored-by: Michael Catanzaro <mcatanzaro@redhat.com>