Fix terminology in Ephy.formControlsAssociated
This seriously confused me for a while. We get an array of form
elements, one of which may be a form. It's not an array of forms.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1495>


(cherry picked from commit d110a253f064147e6124d131c40bdacb1d30e7f3)

Co-authored-by: Michael Catanzaro <mcatanzaro@redhat.com>