ci: suppress dead code warnings from scan-build
There is a strange false positive warning in generated code:

[228/449] Compiling C object src/epiphany-search-provider.p/meson-generated_.._ephy-shell-search-provider-generated.c.o
src/ephy-shell-search-provider-generated.c:2073:37: warning: Value stored to 'skeleton' during its initialization is never read [deadcode.DeadStores]
 2073 |   EphyShellSearchProvider2Skeleton *skeleton = EPHY_SHELL_SEARCH_PROVIDER2_SKELETON (_skeleton);
      |                                     ^~~~~~~~   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
1 warning generated.

But the variable is actually used, twice.

We should find a better way to do this without suppressing the warning
for the entire project, but this will do for now to unblock development.