location-controller: Fix crash when destroying window
When the EphyWindow is destroyed, EphyLocationEntry loses focus and
emits its get-location signal. There is no longer any active embed, so
we shouldn't expect one. This happens after resizing the window while a
JavaScript alert dialog is displayed, which due to a different bug
sometimes causes the web process to hang. It can also happen when
opening Epiphany with just one tab and then immediately pressing Ctrl+W.

There are probably other ways to fix this, but this is the simple
approach.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1516>