Stop passing focus stealing prevention timestamps around
We go to a pretty considerable effort to not drop the focus stealing
prevention timestamp so that we can eventually pass it to
gtk_window_present_with_time() in exactly one place in ephy-shell.c.
This function has been deprecated in favor of gtk_window_present(), so
stop trying.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1519>