shell: fix crash when starting secondary instance
It's an error to call g_variant_builder_end() when the type is still
indefinite. It's indefinite because we have told it to create an array
but it doesn't know what the array elements are until we add at least
one. If we don't have any elements to add, don't send the startup
context at all.

Fixes #2384

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1526>