Remove run-tests from flatpak-builder manifest
This cannot work because our tests rely on a headless compositor or Xvfb
and also on a D-Bus session bus, but we don't have any way to set that
up from flatpak-builder. It's better to run tests on CI or from a
non-flatpak build instead.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1534>