Fix parsing of icon relative paths in web app manifest
If the web app icon is a relative path, we append the path to the
current page's URL, creating a path that looks like e.g.:

https://gitlab.gnome.org/GNOME/epiphany/-/issues/2436/-/pwa-icons/logo-512.png

from example path /-/pwa-icons/logo-512.png. But the path is supposed to
be relative to the manifest URL, not the current page.

Fixes #2436

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1613>