web-view: stop touching the background color
This was added to improve dark mode by avoiding flashes of light
background before loading dark pages. Unfortunately it doesn't work. The
call to reset the background color in LOAD_COMMITTED stage is racy: it
may happen after the website has already set its own background color,
improperly overriding a dark background color to white. This is surely
worse than the flash of light that this code is trying to fix. And there
is nowhere else we can plausibly reset the background color, so the best
we can do is give up. This will need to be fixed in WebKit.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1630>