privacy-report: Move the view switchers to the headerbar
The AdwViewStack contans an AdwPreferencesGroup. However,
the AdwViewStack is already inside an AdwPreferencesGroup.

To avoid nested widgets, move the view switchers to the
headerbar.

Also, move the "no-trackers-page" below the
"privacy-report-page" in the GtkTemplate file. Doing so,
the visibility of the GtkStackPages and the AdwHeaderBar
title widget will work correctly.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1653>