Move AdwToolbarView into EphyBookmarksDialog file
Moves the AdwToolbarView from the EphyWindow file into the
EphyBookmarksDialog file and UI definition. This is to simplify the code and make further changes involving the toolbar view and bookmarks dialog easier. An AdwBin had to be created with the AdwToolbarView as the child since the parent instance can't be a AdwToolbarView.

Fix incorrect variable type

The user_data variable in the on_close_button_clicked function was
incorrectly set as a gboolean instead of a gpointer. This is now fixed.

Remove unnecessary variable

The close_button variable was not needed.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1663>