session: fix crash on startup when no tabs are present
If the window is saved in session state, but no tabs are saved, then we
crash on startup. I don't think we have any bug reports for this, so
apparently I'm the first to notice?

Instead of trying to focus a GtkWidget that doesn't exist, let's create
a tab and load the user's preferred homepage there.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1677>