Always send notif when web app install fails
If there was an error in preparing the web app install outside of
Epiphany, only a warning would be printed to the terminal. Now, a
notification is also sent to the user to inform them of the error.

Closes #2562

Don't transfer ownership of error

Now, the error is freed where it's received.

Use g_autoptr where possible

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1693>