window-commands: print warning when web app installation fails
We have a user who is unable to read the notification due to
ellipsization with a small screen size, and the warning is not printed
to stderr so there's no way to know what is wrong. Let's always print
the warning when something has gone wrong.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1698>