Normalize URLs in the webapp URLs dialog
Previously, user input was saved as typed regardless of whether it is
valid, and this was resolved inside ephy_web_application_is_uri_allowed.
To make it clearer to a user how their input will be interpreted,
normalize URLs when the dialog saves, adding a scheme automatically
based on the webapp's main URL scheme. In addition, report an error and
ignore the user's input if the resulting URL is invalid.