password-export: escape fields that contain commas and quotes
This is a comma-separated value file and the delimiter is the comma. If
we don't escape commas in the output, bad things happen.

Fields containing double quote characters are also currently broken. The
password import code respects RFC 4180, but the export code does not.
The export code already attempts to escape all double quotes with a
second double quote, but when doing so, the entire field needs to be
surrounded in double quotes.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1722>