Remove Granite support from Tech Preview and Canary
This is a test to see if anybody notices.

Granite has recently gained a dependency on shumate, which itself has
dependencies on protobuf and protobuf-c. These are possible to build
from tarballs, but I am hitting errors that I have not yet resolved when
trying to build from git master/main, as we should for a nightly build.
protobuf also takes a fair while to build. I think we can make our CI
significantly faster by just not building the Granite support anymore.

The question is: how many users are testing development versions of
Epiphany on elementary OS? I suspect almost all users are using the
elementary OS or Flathub builds, which do still support Granite, rather
than our development builds. If I'm wrong, please do complain.

This commit notably fixes our CI. Otherwise, we have to add shumate,
since it's a mandatory dependency of Granite, whether we need to display
maps or not. (We'll just have to do this for the Flathub build.)

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1729>