Simplify decision to remember passwords
Instead of keeping the should_remember_passwords GSettings variable and
private_profile boolean variables separately, just decide in the UI
process whether we should remember passwords or not, and send the
decision to the web process.

This notably fixes a minor bug where the password manager JavaScript is
registered even if passwords are not supposed to be remembered.

Also, this removes a comment I wrote claiming that we ought to remember
passwords in incognito mode. This no longer makes sense; the point of
incognito mode is to have separate cookies and local data. Autofilling
passwords without user consent subverts that.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1745>