web-view: fix HTTP auth on authenticationtest.com
Currently we handle HTTP authentication as follows:

 * Asynchronously begin checking password manager for saved credentials
 * Cancel WebKit's default authentication signal handler, so it will
   wait for us instead of displaying an authentication dialog
 * If no credentials are found, provide bogus empty credentials that
   are intended to cause authentication to fail, triggering a retry
 * On retry, let the default signal handler run and show the auth
   dialog.

This fails on https://authenticationtest.com/HTTPAuth/. I think the
website simply does not allow retries because it wants to display an
error message on authentication failure.

Forcing authentication to fail seems pretty overkill. Ideally, Epiphany
would display its own authentication dialog, but that requires work.
Instead, let's just call the default signal handler, which is what we
would have otherwise done anyway if not for the need to handle the
function asynchronously. This is a hack and it would be better to add
new WebKit API to allow this, but it works and I doubt it will break in
the future.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1745>