Don't consider og:image for web app icon
og:image is used for media preview when sharing links on social media.
It's not generally an application icon and is unlikely to be better than
a favicon. For example, on
https://www.cnn.com/politics/live-news/trump-presidency-news-04-29-25/index.html
the og:image is a picture of Donald Trump, which is not an appropriate
choice for web app icon. On https://www.gnome.org the og:image is a
picture of a laptop.

I considered changing the priorities of the icons without removing
og:image completely, since favicons are a bad fallback, but I think they
are better than this.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1755>


(cherry picked from commit 9f359a462491b9da5c0c648451dd944af578eafd)

Co-authored-by: Michael Catanzaro <mcatanzaro@redhat.com>