Update add bookmark dialog
This changes the UI of the bookmark properties dialog when adding a new
bookmark. Note that when creating tags in the add bookmark dialog, the
tag will only be saved if its added to the bookmark, and if the bookmark
is added.

Closes #2640

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1788>