Add function to update tags order without listbox
It's possible that a tags order could get desynced with its actual
contents if changes were made to it while it's not the tag detail tag.
This adds a new function that updates the tags order without it by using the existing variant.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1792>